+++ R/toxicity_generators.R
@@ -12,16 +12,6 @@
 # (c) The scaling of the pharmacological effects of these doses
 # (d) Elaboration of pharmacologic effects (e.g., *graded* toxicities)
 #
-# I have dealt with (1b) in isolation already, by generalizing 'simulate_trials'
-# to the case where 'true_tox_probs' is (effectively) a matrix.
-#
-# A next step may be to handle (2a,b) *together*, in a class which allows for
-# specification of an MTDi distribution over real doses.
-# In constructing this class, I may wish to anticipate the need to abstract
-# the parameter specification to facilitate (1a).
-#
-#' @importFrom distr6 SDistribution Lognormal Rayleigh
-setOldClass(c("Lognormal","SDistribution"))
 
 setClass("mtdi_generator",
   slots = list(
@@ -33,18 +23,18 @@ setClass("mtdi_distribution",
   slots = list(
     CV = "numeric"      # Establish CV and median as universal parameters
   , median = "numeric"  # for specifying MTDi distributions.
-  , dist = "SDistribution"
+  , dist = "list" # think 'distr6::SDistribution'
   ),
   contains = c("mtdi_generator","VIRTUAL")
 )
 
 setGeneric("plot")
 
-#' Visualize K samples from an \code{mtdi_generator} object
+#' Visualize n samples from an \code{mtdi_generator} object
 #'
 #' @param x An \code{mtdi_generator} object
 #' @param y Included for compatibility with generic signature
-#' @param K Number of samples to draw from hyperprior for visualization
+#' @param n Number of samples to draw from hyperprior for visualization
 #' @param col Color of lines used to depict samples
 #' @param \dots Additional arguments passed onward to \code{plot}
 #'
@@ -55,17 +45,17 @@ setGeneric("plot")
 #'                                 ,median_mtd = 5
 #'                                 ,median_sdlog = 0.5
 #'                                 ,units="mg/kg")
-#' plot(mtdi_gen, K=100, col=adjustcolor("red", alpha=0.5))
+#' plot(mtdi_gen, n=100, col=adjustcolor("red", alpha=0.5))
 #' } 
 #' @export
-setMethod("plot", "mtdi_generator", function(x, y=NULL, K=20, col="gray", ...) {
+setMethod("plot", "mtdi_generator", function(x, y=NULL, n=20, col="gray", ...) {
   xlab <- paste0("Dose (", x@units, ")")
   ylab <- "CDF"
   params <- paste0("CV ~ Raleigh(mode=", x@CV, ");  median = "
                    , x@median_mtd, x@units, " \u00b1 " # <-- plus/minus character
                    , 100*x@median_sdlog, "%")
-  mtdi_samples <- draw_samples(x, K = K)
-  title <- paste(K, class(mtdi_samples[[1]]@dist)[1]
+  mtdi_samples <- draw_samples(x, n = n)
+  title <- paste(n, mtdi_samples[[1]]@dist$name
                  , "MTDi distributions sampled from hyperprior")
   CDFs <- seq(0.01, 0.99, 0.01)
   quantiles <- lapply(mtdi_samples
@@ -86,8 +76,8 @@ setMethod("plot", "mtdi_generator", function(x, y=NULL, K=20, col="gray", ...) {
   minor_ticks <- as.vector(outer(2:9, exponents, function(x,y) x*10^y))
   axis(1, at=minor_ticks, tcl=-0.3, labels=NA) # minor ticks
   dose_levels <- getOption('dose_levels')
-  for(k in 2:K){
-    lines(CDFs ~ quantiles[[k]], col = col, ...)
+  for(i in 2:n){
+    lines(CDFs ~ quantiles[[i]], col = col, ...)
   }
   if( !is.null(dose_levels) ){
     abline(v = dose_levels, lty = 3)
@@ -115,7 +105,7 @@ setMethod("plot", "mtdi_generator", function(x, y=NULL, K=20, col="gray", ...) {
 setMethod("plot", "mtdi_distribution", function(x, y=NULL, ...) {
   xlab <- paste0("Dose (", x@units, ")")
   ylab <- "CDF"
-  title <- paste(class(x@dist)[1], "MTDi Distribution")
+  title <- paste(x@dist$name, "MTDi Distribution")
   params <- paste0("CV = ", x@CV, "; median = ", x@median, x@units)
   # I will presume most pharmacology should be done in log-dose space...
   CDFs <- seq(0.01, 0.99, 0.01)
@@ -146,20 +136,27 @@ setMethod("plot", "mtdi_distribution", function(x, y=NULL, ...) {
 
 #' A lognormal MTDi distribution
 #' 
-#' @slot dist An object of class \code{distr6::Lognormal}
+#' @slot dist A list with \code{cdf}, \code{quantile} and \code{name} components
+#'  intended to provide that portion of the interface of \code{distr6::Lognormal}.
 #'
 #' @export mtdi_lognormal
 mtdi_lognormal <- setClass("mtdi_lognormal",
   slots = list(
-    dist = "Lognormal"
+    dist = "list" # think 'distr6::Lognormal'
   ),
   contains = "mtdi_distribution"
 )
 
+#' @importFrom stats qlnorm plnorm
 setMethod("initialize", "mtdi_lognormal",
   function(.Object, CV, median, ...) {
-    .Object@dist <- Lognormal$new(meanlog = log(median),
-                                  sdlog = sqrt(log(CV^2 + 1)))
+    meanlog <- log(median)
+    sdlog <- sqrt(log(CV^2 + 1))
+    .Object@dist <- list(
+      quantile = function(p) qlnorm(p, meanlog = meanlog, sdlog = sdlog)
+    , cdf = function(q) plnorm(q, meanlog = meanlog, sdlog = sdlog)
+    , name = "Lognormal"
+    )
     # Initialize 'bottom-up' to avoid a premature validity check:
     .Object <- callNextMethod(.Object, CV=CV, median=median, ...)
   })
@@ -168,20 +165,27 @@ setClass("hyper_mtdi_distribution",
   contains = c("mtdi_generator","VIRTUAL")
   )
 
-setGeneric("draw_samples", function(hyper, K, ...) {
+setGeneric("draw_samples", function(hyper, n, ...) {
   standardGeneric("draw_samples")
 })
 
 #' Hyperprior for lognormal MTDi distributions
 #' 
 #' This hyperprior generates lognormal MTDi distributions with their
-#' coefficient of variation being drawn from a Rayleigh distribution
-#' with mode parameter \code{CV}. Because the standard deviation of
-#' this distribution is
-#' 
-#' this implicitly link our uncertainty about \code{CV} to its value.
-#' 
-#' The medians of the lognormal distributions generated are themselves
+#' coefficients of variation being drawn from a Rayleigh distribution
+#' with mode parameter set to
+#'  \deqn{\sigma := CV.}
+#' Because the standard deviation of this distribution is
+#'  \deqn{sd = \sigma \sqrt{(2 - \pi/2)} \approx 0.655 \sigma,}{%
+#'        sd = \sigma \sqrt(2 - \pi/2) ~ 0.655 \sigma,}
+#' this conveniently links our uncertainty about \code{CV} to its \emph{mode},
+#' and indeed to other measures of centrality that are proportional to this:
+#'  \deqn{mean = \sigma \sqrt{\pi/2} \approx 1.25 \sigma}{%
+#'        mean = \sigma \sqrt(\pi/2) ~ 1.25 \sigma}
+#'        
+#'  \deqn{median = \sigma \sqrt{2 log(2)} \approx 1.18 \sigma.}{%
+#'        median = \sigma \sqrt(2 log(2)) ~ 1.18 \sigma.}
+#' The \emph{medians} of the lognormal distributions generated are themselves
 #' drawn from a lognormal distribution with \code{meanlog = log(median_mtd)}
 #' and \code{sdlog = median_sdlog}. Thus, parameter \code{median_sdlog}
 #' represents a proportional uncertainty in \code{median_mtd}.
@@ -205,15 +209,16 @@ setMethod("initialize", "hyper_mtdi_lognormal",
     .Object <- callNextMethod(.Object, ...)
   })
 
-#' @importFrom stats rlnorm
+#' @importFrom stats rlnorm rchisq
 setMethod(
   "draw_samples"
   , c("hyper_mtdi_lognormal","numeric"),
-  function(hyper, K=NULL, ...) {
-    if(missing(K)) K <- 1
+  function(hyper, n=NULL, ...) {
+    if(missing(n)) n <- 1
     mapply(mtdi_lognormal
-          , CV = Rayleigh$new(mode = hyper@CV)$rand(n=K, simplify=TRUE)
-          , median = rlnorm(n=K
+          #, CV = Rayleigh$new(mode = hyper@CV)$rand(n=n, simplify=TRUE)
+          , CV = sqrt(rchisq(n=n, df=2))*hyper@CV # Rayleigh(1) is chi dist w/ 2 d.f.
+          , median = rlnorm(n=n
                             , meanlog=log(hyper@median_mtd)
                             , sdlog=hyper@median_sdlog)
           , MoreArgs = list(units = hyper@units
@@ -221,3 +226,14 @@ setMethod(
           , SIMPLIFY = FALSE)
   })
 
+#' Test performance of \code{draw_samples} function
+#' 
+#' @param n Number of samples to draw
+#'
+#' @export
+test_draw_samples <- function(n = 500){
+  mtdi_gen <- hyper_mtdi_lognormal(CV = 1
+                                   , median_mtd = 6, median_sdlog = 0.5
+                                   , units="mg/kg")
+  invisible(draw_samples(mtdi_gen, n = n))
+}